moves getApiUrl from api plugin into a helper function.
